requirements.txt
gitdb==4.0.5
GitPython==3.1.3
Jinja2==2.11.2
MarkupSafe==1.1.1
smmap==3.0.4
